Script monitor splitting
Monitor splitting has been moved to a script that will only split the
monitor if it supports a resolution of 5120 by 1440. This roughly
enables support for different environments.

There is one caveat in regards to i3wm. Workspaces will properly
fallback to any given monitor, but three status bars will still be
created. A possible solution could be generating the configuration file
on boot as there does not appear to be a way internal to i3wm to
conditionally display a status bar.